samba: tag release samba-4.1.22

Conflicts:
	.gitignore
	VERSION
	WHATSNEW.txt
	buildtools/wafsamba/samba_patterns.py
	buildtools/wafsamba/samba_python.py
	buildtools/wafsamba/samba_utils.py
	docs-xml/Samba3-HOWTO/TOSHARG-Diagnosis.xml
	docs-xml/build/DTD/samba.entities
	docs-xml/manpages/dbwrap_tool.1.xml
	docs-xml/manpages/samba-tool.8.xml
	docs-xml/manpages/samba.7.xml
	docs-xml/manpages/samba.8.xml
	docs-xml/manpages/vfs_catia.8.xml
	docs-xml/manpages/winbindd.8.xml
	docs-xml/smbdotconf/protocol/aclallowexecutealways.xml
	docs-xml/smbdotconf/protocol/clientminprotocol.xml
	docs-xml/smbdotconf/protocol/serverminprotocol.xml
	docs-xml/smbdotconf/vfs/ntvfshandler.xml
	docs-xml/wscript_build
	dynconfig/config.m4
	examples/VFS/skel_opaque.c
	examples/VFS/skel_transparent.c
	extra/smbprint
	file_server/file_server.c
	lib/ccan/libccan.m4
	lib/ldb/ldb_tdb/ldb_tdb.c
	lib/ldb/wscript
	lib/replace/getpass.c
	lib/replace/libreplace.m4
	lib/replace/libreplace_macros.m4
	lib/replace/replace-testsuite.h
	lib/replace/wscript
	lib/tdb/common/io.c
	lib/tdb/common/open.c
	lib/tdb/common/rescue.c
	lib/tdb/libtdb.m4
	lib/tdb/tools/tdbdump.c
	lib/tdb/wscript
	lib/tevent/libtevent.m4
	lib/tevent/testsuite.c
	lib/tevent/tevent_poll.c
	lib/util/genrand.c
	lib/util/samba_util.h
	lib/util/wscript_build
	libcli/auth/credentials.c
	libcli/ldap/ldap_message.c
	libcli/security/secdesc.c
	librpc/rpc/binding.c
	librpc/wscript_build
	nsswitch/wbinfo.c
	packaging/RHEL-CTDB/configure.rpm
	packaging/RHEL-CTDB/setup/smbprint
	packaging/RHEL/samba.spec.tmpl
	packaging/RHEL/setup/smbprint
	packaging/printing/smbprint
	packaging/systemd/nmb.service
	packaging/systemd/samba.conf.tmp
	packaging/systemd/samba.service
	packaging/systemd/smb.service
	packaging/systemd/winbind.service
	python/samba/dbchecker.py
	python/samba/netcmd/domain.py
	python/samba/netcmd/processes.py
	python/samba/netcmd/user.py
	python/samba/ntacls.py
	python/samba/provision/__init__.py
	python/samba/provision/sambadns.py
	python/samba/samba3/__init__.py
	python/samba/tests/posixacl.py
	python/samba/tests/samba3sam.py
	python/samba/upgrade.py
	script/autobuild.py
	selftest/knownfail
	selftest/target/Samba3.pm
	selftest/target/Samba4.pm
	selftest/tests.py
	source3/Makefile-smbtorture
	source3/Makefile.in
	source3/auth/token_util.c
	source3/configure.in
	source3/include/client.h
	source3/include/nt_printing.h
	source3/include/passdb.h
	source3/include/smb.h
	source3/include/vfs.h
	source3/include/vfs_macros.h
	source3/lib/sysacls.c
	source3/lib/util_sid_passdb.c
	source3/libads/kerberos.c
	source3/libads/kerberos_keytab.c
	source3/libnet/libnet_join.c
	source3/librpc/idl/open_files.idl
	source3/libsmb/clidfs.c
	source3/locking/proto.h
	source3/locking/share_mode_lock.c
	source3/m4/check_path.m4
	source3/modules/nfs4_acls.c
	source3/modules/vfs_acl_common.c
	source3/modules/vfs_aixacl2.c
	source3/modules/vfs_default.c
	source3/modules/vfs_dirsort.c
	source3/modules/vfs_fake_acls.c
	source3/modules/vfs_gpfs.c
	source3/modules/vfs_irixacl.c
	source3/modules/vfs_shadow_copy2.c
	source3/modules/vfs_streams_depot.c
	source3/modules/vfs_zfsacl.c
	source3/modules/wscript_build
	source3/passdb/pdb_interface.c
	source3/passdb/pdb_ldap.c
	source3/printing/nt_printing_ads.c
	source3/rpc_server/netlogon/srv_netlog_nt.c
	source3/rpc_server/spoolss/srv_spoolss_nt.c
	source3/script/installman.sh
	source3/script/tests/test_smbclient_s3.sh
	source3/selftest/s3-selftest.sh
	source3/selftest/tests.py
	source3/smbd/close.c
	source3/smbd/dir.c
	source3/smbd/durable.c
	source3/smbd/globals.c
	source3/smbd/msdfs.c
	source3/smbd/nttrans.c
	source3/smbd/open.c
	source3/smbd/posix_acls.c
	source3/smbd/pysmbd.c
	source3/smbd/reply.c
	source3/smbd/server.c
	source3/smbd/service.c
	source3/smbd/smb2_create.c
	source3/smbd/smb2_ioctl.c
	source3/smbd/smb2_server.c
	source3/smbd/smb2_sesssetup.c
	source3/smbd/smbXsrv_session.c
	source3/smbd/smbXsrv_tcon.c
	source3/smbd/uid.c
	source3/torture/test_cleanup.c
	source3/torture/torture.c
	source3/utils/net_dns.c
	source3/utils/net_rpc.c
	source3/utils/net_serverid.c
	source3/utils/smbcontrol.c
	source3/web/cgi.c
	source3/web/swat.c
	source3/web/swat_proto.h
	source3/winbindd/idmap_autorid.c
	source3/winbindd/idmap_util.c
	source3/winbindd/wb_getgrsid.c
	source3/winbindd/winbindd.c
	source3/winbindd/winbindd_ads.c
	source3/winbindd/winbindd_cm.c
	source3/winbindd/winbindd_pam.c
	source3/winbindd/winbindd_util.c
	source3/winbindd/wscript_build
	source3/wscript
	source3/wscript_build
	source4/dns_server/dns_utils.c
	source4/dsdb/common/util.c
	source4/dsdb/samdb/ldb_modules/acl_read.c
	source4/dsdb/samdb/ldb_modules/descriptor.c
	source4/dsdb/samdb/ldb_modules/extended_dn_in.c
	source4/dsdb/samdb/ldb_modules/linked_attributes.c
	source4/dsdb/samdb/ldb_modules/operational.c
	source4/dsdb/samdb/ldb_modules/repl_meta_data.c
	source4/dsdb/samdb/ldb_modules/rootdse.c
	source4/dsdb/samdb/ldb_modules/samldb.c
	source4/dsdb/samdb/ldb_modules/schema_data.c
	source4/dsdb/samdb/ldb_modules/secrets_tdb_sync.c
	source4/dsdb/schema/schema_inferiors.c
	source4/dsdb/schema/schema_set.c
	source4/dsdb/tests/python/ldap.py
	source4/lib/cmdline/wscript_build
	source4/libcli/finddcs_cldap.c
	source4/ntp_signd/ntp_signd.c
	source4/rpc_server/dcerpc_server.c
	source4/rpc_server/drsuapi/getncchanges.c
	source4/rpc_server/netlogon/dcerpc_netlogon.c
	source4/scripting/bin/samba_dnsupdate
	source4/scripting/devel/crackname
	source4/selftest/tests.py
	source4/smbd/server.c
	source4/torture/dns/dlz_bind9.c
	source4/torture/rpc/netlogon.c
	source4/torture/smb2/compound.c
	source4/torture/smb2/notify.c
	swat/lang/ru/help/welcome-no-samba-doc.html
	swat/lang/ru/help/welcome.html
	swat/lang/ru/images/globals.gif
	swat/lang/ru/images/home.gif
	swat/lang/ru/images/passwd.gif
	swat/lang/ru/images/printers.gif
	swat/lang/ru/images/shares.gif
	swat/lang/ru/images/status.gif
	swat/lang/ru/images/viewconfig.gif
	swat/lang/ru/images/wizard.gif
	swat/lang/ru/include/header.html
	testprogs/blackbox/dbcheck-oldrelease.sh
	testprogs/blackbox/dbcheck.sh
	wscript

+++ source3/printing/printing.c
@@ -197,13 +197,18 @@ bool print_backend_init(struct messaging_context *msg_ctx)
 	const char *sversion = "INFO/version";
 	int services = lp_numservices();
 	int snum;
+	bool ok;
 
 	if (!printer_list_parent_init()) {
 		return false;
 	}
 
+	ok = directory_create_or_exist(cache_path("printing"), geteuid(), 0755);
+	if (!ok) {
+		return false;
+	}
+
 	unlink(cache_path("printing.tdb"));
-	mkdir(cache_path("printing"),0755);
 
 	/* handle a Samba upgrade */
 
@@ -2738,6 +2743,7 @@ static WERROR print_job_spool_file(int snum, uint32_t jobid,
 	SMB_STRUCT_STAT st;
 	const char *path;
 	int len;
+	mode_t mask;
 
 	/* if this file is within the printer path, it means that smbd
 	 * is spooling it and will pass us control when it is finished.
@@ -2762,7 +2768,7 @@ static WERROR print_job_spool_file(int snum, uint32_t jobid,
 			fstrcpy(pjob->filename, output_file);
 
 			DEBUG(3, ("print_job_spool_file:"
-				  "External spooling activated"));
+				  "External spooling activated\n"));
 
 			/* we do not open the file until spooling is done */
 			pjob->fd = -1;
@@ -2775,7 +2781,9 @@ static WERROR print_job_spool_file(int snum, uint32_t jobid,
 	slprintf(pjob->filename, sizeof(pjob->filename)-1,
 		 "%s/%sXXXXXX", lp_pathname(talloc_tos(), snum),
 		 PRINT_SPOOL_PREFIX);
+	mask = umask(S_IRWXO | S_IRWXG);
 	pjob->fd = mkstemp(pjob->filename);
+	umask(mask);
 
 	if (pjob->fd == -1) {
 		werr = map_werror_from_unix(errno);
